requirements.txt
streamlit
pandas
pymongo
matplotlib
openai
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-qdrant
langchain-text-splitters
qdrant-client
Pillow
easyocr
sentence-transformers
pypdf
streamlit-authenticator
requests
plotly
asyncio